Changing default value for fpsInVps to TRUE (#222)
This change was result of investigating issues/218 where ffprobe did not report the correct frame rate.  The solution was to use the fpsInVps command line option.  However a better solution is to change the default so that the fps is always written to the Vps structure (unless the user decides to set fpsInVps to false with a parameter).