SpatialFullDistortionKernel(SSE) fix (#195)
* SpatialFullDistortionKernel(SSE)

Fix bug in function where subtraction of 8 bit unsigned integers was landing in an 8 bit signed space (not enough bits to store possible answers).  Propsed fix to use zero extend or unpack so that subtraction of 8 bit unsigned integers land in 16 bit space ( leaving enough bits to store possible answers).

* SadLoopKernel(AVX2) and SadLoopKernelHme0(AVX2/AVX512)

Fix bug where the loading of 4 data rows was done in an incorrect order, only in the case where width==4 and height%4=0.  Instead of loading (3, 1) and (2, 4), we change the behavior to load (1, 3) and (2, 4) so that mpsad with the source can be done correctly.  This order matches what is done in other places inside of SadLoopKernel, e.g. the case where width==8 and height%4=0.

* Compute40xMSad(AVX2) and Compute56xMSad(AVX2)

Certain input sizes compute Sad with 56 width sizes.  Add 2 new functions in ASM code so that Sad can be computed for 40 and 56 width sizes to mantain compatibility with C code in these cases.